events: security stuff

+++ src/events/messageReactionAdd.js
@@ -14,16 +14,21 @@ module.exports = class Event {
 		// in a dev environment this will let people re-run commands
 		if (this.Atlas.env === 'development' && emoji.name === '🔁') {
 			const fetchedMsg = await this.Atlas.client.getMessage(msg.channel.id, msg.id);
-			this.Atlas.client.emit('messageCreate', fetchedMsg);
-			this.Atlas.client.removeMessageReaction(
-				fetchedMsg.channel.id,
-				fetchedMsg.id,
-				'🔁',
-				userID,
-			)
-				.catch(() => false);
-			this.Atlas.client.purgeChannel(msg.channel.id, 100, m => m.author.bot, null, fetchedMsg.id)
-				.catch(() => false);
+
+			if (userID === fetchedMsg.author.id) {
+				try {
+					await this.Atlas.client.purgeChannel(msg.channel.id, 100, m => m.author.bot, null, fetchedMsg.id);
+
+					await this.Atlas.client.removeMessageReaction(
+						fetchedMsg.channel.id,
+						fetchedMsg.id,
+						'🔁',
+						userID,
+					);
+				} catch (e) {} // eslint-disable-line no-empty
+
+				this.Atlas.client.emit('messageCreate', fetchedMsg);
+			}
 		}
 	}
 };
